change: プロダクトコード側のファイルが `authOptions` と `route` に分かれたためテストコード側も分ける
